refactor: isolate mood module
