+++ AGENTS.md
@@ -1,0 +1,4 @@
+# Coding Guidelines
+- Keep domain features under `com.hab.blog.module` packages to reduce coupling.
+- When adding new modules, group controller, service, repository and entity classes together inside the module package.
+- Always run `./gradlew test` before committing.
