+++ src/main/java/com/hab/blog/module/mood/Entity/Mood.java
@@ -1,4 +1,4 @@
-package com.hab.blog.api.v1.moods.Entity;
+package com.hab.blog.module.mood.Entity;
 
 import com.hab.blog.api.v1.auth.Entity.User;
 import jakarta.persistence.*;
